ADD MISSION COMPONENT
